+++ packages/sui-studio/bin/sui-studio.js
@@ -28,6 +28,8 @@ program
     }).then(process.exit, process.exit)
   })
 
+program.command('dev <component>', 'Develop an isolate component').alias('d')
+
 program
   .command(
     'generate <category> <component>',