+++ packages/sui-studio/src/components/demo/index.js
@@ -15,39 +15,21 @@ import EventsButtons from './EventsButtons'
 import ThemesButtons from './ThemesButtons'
 import withContext from './HoC/withContext'
 import withProvider from './HoC/withProvider'
-import deepmerge from 'deepmerge'
 
 import {createStore} from '@s-ui/react-domain-connector'
 
-const DEFAULT_CONTEXT = 'default'
+import {
+  createContextByType,
+  checkIfPackageHasProvider,
+  isFunction,
+  cleanDisplayName,
+  pipe,
+  removeDefaultContext
+} from './utilities'
+
 const EVIL_HACK_TO_RERENDER_AFTER_CHANGE = ' '
-const DDD_REACT_REDUX = '@schibstedspain/ddd-react-redux'
-const REACT_DOMAIN_CONNECTOR = '@s-ui/react-domain-connector'
 const CONTAINER_CLASS = 'sui-Studio'
 const FULLSCREEN_CLASS = 'sui-Studio--fullscreen'
-
-const createContextByType = (ctxt, type) => {
-  // check if the user has created a context.js with the needed contextTypes
-  if (typeof ctxt !== 'object' || ctxt === null) {
-    console.warn(
-      "[Studio] You're trying to use a contextType in your component but it seems that you haven't created a context.js in the playground folder. This will likely make your component won't work as expected or it might have an useless context."
-    )
-  }
-  return deepmerge(ctxt[DEFAULT_CONTEXT], ctxt[type])
-}
-
-const isFunction = fnc => !!(fnc && fnc.constructor && fnc.call && fnc.apply)
-const cleanDisplayName = displayName => {
-  const [fallback, name] = displayName.split(/\w+\((\w+)\)/)
-  return name !== undefined ? name : fallback
-}
-const pipe = (...funcs) => arg =>
-  funcs.reduce((value, func) => func(value), arg)
-
-const removeDefaultContext = exports => {
-  const {[DEFAULT_CONTEXT]: toOmit, ...restOfExports} = exports
-  return restOfExports
-}
 
 export default class Demo extends Component {
   static async bootstrapWith(demo, {category, component, style, themes}) {
@@ -104,15 +86,6 @@ export default class Demo extends Component {
       const themes = themesFor({category, component})
       Demo.bootstrapWith(this, {category, component, style, themes})
     })
-  }
-
-  _checkIfPackageHasProvider({pkg}) {
-    return (
-      pkg &&
-      pkg.dependencies &&
-      (pkg.dependencies[DDD_REACT_REDUX] ||
-        pkg.dependencies[REACT_DOMAIN_CONNECTOR])
-    )
   }
 
   componentDidMount() {
@@ -189,7 +162,7 @@ export default class Demo extends Component {
     const contextTypes = Base.contextTypes || Base.originalContextTypes
     const context = contextTypes && createContextByType(ctxt, ctxtType)
     const {domain} = context || {}
-    const hasProvider = this._checkIfPackageHasProvider({pkg})
+    const hasProvider = checkIfPackageHasProvider(pkg)
     const store = domain && hasProvider && createStore(domain)
 
     const Enhance = pipe(