+++ packages/sui-studio/src/components/preview/index.js
@@ -8,13 +8,11 @@ const ERROR_TIMEOUT = 500
 
 export default class Preview extends Component {
   static propTypes = {
-    className: PropTypes.string,
     code: PropTypes.string,
     scope: PropTypes.object
   }
 
   static defaultProps = {
-    className: '',
     scope: {React}
   }
 
@@ -28,9 +26,7 @@ export default class Preview extends Component {
 
   componentDidUpdate(prevProps) {
     clearTimeout(this.timeoutID)
-    if (this.props.code !== prevProps.code) {
-      this.executeCode()
-    }
+    this.executeCode()
   }
 
   setTimeout() {